feat(dashboard): implemented tabs to dashboard admin views
