--- src/components/common/TabPanelGeneric.jsx
@@ -1,31 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { Box, Typography } from "@mui/material";
-
-function TabPanelGenric(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <Box
-      role="tabpanel"
-      hidden={value !== index}
-      id={`full-width-tabpanel-${index}`}
-      aria-labelledby={`full-width-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box sx={{ p: 3 }}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </Box>
-  );
-}
-
-TabPanelGenric.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.number.isRequired,
-  value: PropTypes.number.isRequired,
-};
-
-export default TabPanelGenric;
